requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
disnake==2.8.0
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
mathparse==0.1.2
multidict==6.0.4
nltk==3.8.1
Pint==0.22
pymongo==3.13.0
python-dateutil==2.7.5
python-dotenv==1.0.0
pytz==2023.3
PyYAML==3.13
regex==2023.6.3
six==1.16.0
SQLAlchemy==1.2.19
tqdm==4.65.0
typing_extensions==4.6.3
yarl==1.9.2
